finish

+++ src/app/glsl/colorShader_fs.frag
@@ -1,4 +1,4 @@
-varying mediump vec4 vColor;
+varying highp vec4 vColor;
 
 void main() {
     gl_FragColor = vColor;
